django/requirements.txt
django
djangorestframework
black
django-phonenumber-field
phonenumbers
coverage